+++ src/UX/ContractCreator/AppSelector.css
