+++ src/UX/ContractCreator/CreateContractModal.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
-import { Menu } from '../Compiler/Menu';
+import { Menu } from './Menu';
 
 export class CreateContractModal extends React.Component {
     constructor(props) {
